requirements.txt
discord.py==2.3.2
PyNaCl==1.5.0
yandex-music==2.1.1
yt-dlp==2023.12.30
youtube-search-python==1.6.6
python-dotenv==1.0.0
